update : 공통 레이아웃 footer부분 보완
