+++ src/components/layout/CommonLayout/components/FloatingFooter/FloatingFooter.styles.ts
