+++ src/components/layout/CommonLayout/components/FloatingFooter/FloatingFooter.tsx
@@ -1,0 +1,18 @@
+import React from 'react';
+import {
+  StFloatingFooter,
+  StFloatingFooterFixed,
+} from './FloatingFooter.styles';
+interface FooterProps {
+  children: React.ReactNode;
+}
+
+export const FloatingFooter = ({ children }: FooterProps) => {
+  return (
+    <StFloatingFooter>
+      <StFloatingFooterFixed className="fixed">
+        {children}
+      </StFloatingFooterFixed>
+    </StFloatingFooter>
+  );
+};
