+++ src/components/layout/index.ts
@@ -1,3 +1,2 @@
 export * from './NavBar/NavBar';
 export * from './CommonLayout/CommonLayout';
-export * from './FloatingFooter/FloatingFooter';
